pyproject.toml
[tool.poetry]
name = "automodeling"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
jupyterlab = "^4.0.7"
langchain = "^0.0.329"
llama-index = "^0.8.56"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
botocore = "1.31.76"
boto3 = "1.28.76"
boto3-stubs = { version = "1.28.76", extras = ["essential"] }
pdftotext = "^2.2.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"
black = "^23.10.1"

[tool.poetry.group.standalone]
optional = true

[tool.poetry.group.standalone.dependencies]
apache-airflow = "^2.7.2"
connexion = "^2.14.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
